pyproject.toml
[tool.poetry]
name = "flaskblog"
version = "0.1.0"
description = "Flask API that allow create articles"
authors = ["HolyDorus <marichevdenysmail@gmail.com> <https://github.com/HolyDorus>"]

[tool.poetry.dependencies]
python = "^3.8"
flask = "^1.1.2"
Flask-SQLAlchemy = "^2.4.4"
psycopg2 = "^2.8.5"
pyjwt = "^1.7.1"
python-dotenv = "^0.14.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
